Move test setup helpers into shared support with proper mixin config
